requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-import-export==3.3.7
django-jalali==6.0.1
django-jalali-date==1.1.0
django-sql-explorer==4.0.2
et-xmlfile==1.1.0
gunicorn==21.2.0
holidays==0.42
jdatetime==4.1.1
MarkupPy==1.14
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tzdata==2024.1
xlrd==2.0.1
xlwt==1.3.0
